Move request classes to Requestors namespace
