+++ src/Authenticator.php
@@ -15,8 +15,8 @@ use Pinnacle\OpenIdConnect\Models\AuthorizationCodeCallbackData;
 use Pinnacle\OpenIdConnect\Models\AuthorizationCodeResponse;
 use Pinnacle\OpenIdConnect\Models\UserInfo;
 use Pinnacle\OpenIdConnect\Provider\Contracts\ProviderConfigurationInterface;
-use Pinnacle\OpenIdConnect\Services\RequestUserInfo;
-use Pinnacle\OpenIdConnect\Services\TokenRequestor;
+use Pinnacle\OpenIdConnect\Requestors\RequestUserInfo;
+use Pinnacle\OpenIdConnect\Requestors\TokenRequestor;
 use Pinnacle\OpenIdConnect\State\Contracts\StatePersisterInterface;
 use Pinnacle\OpenIdConnect\State\Exceptions\StatePersisterMissingValueException;
 use Pinnacle\OpenIdConnect\State\StatePersisterWrapper;
